libdrm: Store bo offset as 64-bit

+++ tegra/private.h
@@ -41,7 +41,7 @@ struct drm_tegra {
 struct drm_tegra_bo {
 	struct drm_tegra *drm;
 	uint32_t handle;
-	uint32_t offset;
+	uint64_t offset;
 	uint32_t flags;
 	uint32_t size;
 	atomic_t ref;
